수정: 그룹서비스 인터페이스화
