+++ src/main/kotlin/com/teamsparta/moamoa/event/DiscountPaymentEventListener.kt
@@ -1,15 +1,15 @@
 package com.teamsparta.moamoa.event
 
-import com.teamsparta.moamoa.domain.groupPurchase.service.GroupPurchaseService
+import com.teamsparta.moamoa.domain.groupPurchase.service.GroupPurchaseServiceImpl
 import org.springframework.context.event.EventListener
 import org.springframework.scheduling.annotation.Async
 import org.springframework.stereotype.Component
 
 @Component
-class DiscountPaymentEventListener(private val groupPurchaseService: GroupPurchaseService) {
+class DiscountPaymentEventListener(private val groupPurchaseServiceImpl: GroupPurchaseServiceImpl) {
     @Async
     @EventListener
     fun handleDiscountAppliedEvent(event: DiscountPaymentEvent) {
-        groupPurchaseService.createAndJoinOrJoinGroupPurchase(event.orderId, event.userId)
+        groupPurchaseServiceImpl.createAndJoinOrJoinGroupPurchase(event.orderId, event.userId)
     }
 }
